// src/model/moment.ts
export class Moment {
    // id: number;
    content: string;
    date: Date;
    question: string;
    optionA: string;
    optionB: string;
    optionC: string;
    answer: string;
    author: string = "Jack";
    // readable: boolean = true;
    // doublecheck: boolean;
    icon: string = "md-checkmark";
    cooling: boolean = false;
}